requirements.txt
Flask-Cors==3.0.10
black==23.3.0
pylint==2.17.4
pytest-cov==4.0.0
